+++ shepherd_core/shepherd_core/data_models/base/shepherd.py
@@ -10,7 +10,10 @@ from typing import final
 import ryaml
 from pydantic import BaseModel
 from pydantic import ConfigDict
+from pydantic import model_validator
 from typing_extensions import Self
+
+from shepherd_core.logger import log
 
 from .timezone import local_now
 from .wrapper import Wrapper
@@ -49,7 +52,7 @@ class ShpModel(BaseModel):
 
     model_config = ConfigDict(
         frozen=True,  # -> const after creation, hashable! but currently manually with .get_hash()
-        extra="forbid",  # no unnamed attributes allowed
+        extra="ignore",  # additional parameters get ignored (validator below is warning about them)
         validate_default=True,
         validate_assignment=True,  # not relevant for the frozen model
         str_min_length=1,  # force more meaningful descriptors,
@@ -98,6 +101,18 @@ class ShpModel(BaseModel):
         """Fn of dict."""
         for key in self.keys():
             yield key, self.__getattribute__(key)
+
+    @model_validator(mode="before")
+    @classmethod
+    def __alert_extra_field__(cls, values: dict[str, Any]) -> dict[str, Any]:
+        if isinstance(values, dict) and (
+            extra_fields := values.keys()
+            - cls.model_fields.keys()
+            - {v.alias for v in cls.model_fields.values()}
+        ):
+            log.warning("%s is ignoring extra fields: %s", cls.__name__, extra_fields)
+
+        return values
 
     @final
     @classmethod